tests: pull LooperRunenr out into TestingUtils package for reuse
