+++ sdk/src/androidTest/java/com/punchthrough/bean/sdk/TestBeanManager.java
@@ -1,10 +1,12 @@
 package com.punchthrough.bean.sdk;
 
-import android.os.Looper;
 import android.test.AndroidTestCase;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.CountDownLatch;
+
+import com.punchthrough.bean.sdk.TestingUtils.LooperRunner;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -14,36 +16,46 @@ import static org.assertj.core.api.Assertions.assertThat;
  */
 public class TestBeanManager extends AndroidTestCase {
 
+    private LooperRunner lr = new LooperRunner();
+    private Thread lrThread = new Thread(lr);
     private BeanManager bm;
 
     protected void setUp() {
         this.bm = BeanManager.getInstance();
+        lrThread.start();
     }
 
     protected void tearDown() {
+        /*
+        * "Speaking very generally, never quit() your looper threads. That method exists
+        * mostly for historical and testing reasons. In Real Life™, I recommend that you
+        * continue to reuse the same looper thread(s) for the life of the process rather
+        * than creating/quitting them."
+        * */
+//        lr.quit();
+//        lrThread.join();
     }
 
     public void testDiscovery() throws InterruptedException {
-        final Looper looper = this.bm.mHandler.getLooper();
-
+        final CountDownLatch beanLatch = new CountDownLatch(1);
         final List<Bean> beans = new ArrayList<>();
 
         BeanDiscoveryListener listener = new BeanDiscoveryListener() {
             @Override
             public void onBeanDiscovered(Bean bean, int rssi) {
                 beans.add(bean);
-                looper.quit();
+                beanLatch.countDown();
             }
 
             @Override
             public void onDiscoveryComplete() {
-                looper.quit();
+                beanLatch.countDown();
             }
         };
 
-        boolean startedOK = BeanManager.getInstance().startDiscovery(listener);
+        boolean startedOK = this.bm.startDiscovery(listener);
         assertThat(startedOK).isTrue();
-        looper.loop();
+        beanLatch.await();
         assertThat(beans.size()).isGreaterThan(0);
     }
 }
